# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker API",
    "version": "1.0.0",
    "description": "API for managing workouts and exercises."
  },
  "paths": {
    "/api/workouts": {
      "get": {
        "summary": "Get all workouts",
        "responses": {
          "200": {
            "description": "A list of workouts."
          }
        }
      },
      "post": {
        "summary": "Create a new workout",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Workout created successfully."
          }
        }
      }
    },
    "/api/workouts/{workout_id}/exercises": {
      "get": {
        "summary": "Get all exercises for a workout",
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of exercises."
          }
        }
      },
      "post": {
        "summary": "Create a new exercise for a workout",
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sets",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "reps",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "duration",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Exercise created successfully."
          }
        }
      }
    },
    "/api/exercises/{exercise_id}/complete": {
      "put": {
        "summary": "Toggle the completion status of an exercise",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "completed",
            "in": "body",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise completion status updated."
          }
        }
      }
    },
    "/api/timer": {
      "post": {
        "summary": "Start a timer",
        "parameters": [
          {
            "name": "duration",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timer started successfully."
          }
        }
      }
    }
  }
}